Vary futility margin according to ply
More aggressive pruning is doing far better at STC than at LTC, so for longer searchs use a higher futility margin near the leaves.

Intuitively this makes sense to me, as futility is a depth/accuracy trade-off. At lower depths it's more important to get to the next ply quickly, whereas at higher depths it may be more important to be accurate.

Also has the consequence of a little more pruning in non-PV nodes as the seldepth is lower.

Bench: 5183374